api/requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.9
PyYAML==6.0.2
